refactor: moved locators to Page added documentation
